Moved jquery test from simple to a specific test file
